+++ src/main/webapp/ui/src/util/styles.js.flow
@@ -1,0 +1,13 @@
+// @flow strict
+
+import { type ComponentType } from "react";
+import { type Theme } from "../theme";
+
+declare export function withStyles<Config, Classes: { ... }>({[ Class in keyof Classes]: { ... }}
+    | ((Theme, Config) => {[ Class in keyof Classes]: { ... }})): (ComponentType<{| ...Config, classes: Classes |}>) => ComponentType<Config>;
+
+export type Sx =
+  | { [string]: string | number | ((Theme) => string | number) | Sx }
+  | ((Theme) => { ... } | Array<boolean | { ... } | ((Theme) => { ... })>);
+
+declare export function mergeThemes({ ... }, { ... }): { ... };
